Add first brick of the connector_jira addon
Add project binding

Add export of projects

Add Jira project tasks model

Add external dependencies in manifest

Add security accesses

Add basic import of tasks

Task import are triggered by Jira webhooks. A batch import must be added
later to import the changes from missed hooks or for those not wanting
to use the webhooks.

Add a basic backend adapter

Batch import of project tasks

Remove markdown

The fields are not in markdown, but in a wiki syntax (proper to JIRA yet
be determined)

Link users and import tasks assignees

Import worklogs

Use an adapter for the project export

Import issue types

Filter type of issues to synchronize

And assign the worklog to the first parent we find that we synchronize.

If the type of the worklog's issue is synchronized, the worklog is
assigned on this issue. If the worklog's issue is a sub-task, and the
parent task of the sub-task is of a synchronized type, it link it to it.
If the task has an epic, and the epics are synchronized, it link it to
it. In last resort, the worklog will be linked with the project but with
no task.

Configure the name of the Epic Link field from API

Create webhooks in JIRA from Odoo

Choose template on project creation

Discard Epics from different projects

JIRA allow to choose an epic for any project. Setting the Epic as the
Odoo task would create inconsistencies between the project and the task.
We discard such an Epic, likely the task will be empty.

Delete worklogs deleted on JIRA (only with webhooks)

Delete tasks deleted on JIRA (only with webhooks)

Add JIRA issue type in tasks

Add JIRA Issue Key

Add core business project types

Improve display of JIRA backend

Use a dedicated button to activate the Epic Link.

Link the Epic in a dedicated field

Add JIRA Parent for subtasks

Show JIRA key in name

Create project from shared project

Check JIRA project keys

Remove the button to 'one-click-export' a project

It can't be working because we have other fields to configure on the
binding.

Add help on issue type synchronizations

 Accept tasks without parent

Do not use id builtin