Filter type of issues to synchronize
And assign the worklog to the first parent we find that we synchronize.

If the type of the worklog's issue is synchronized, the worklog is
assigned on this issue. If the worklog's issue is a sub-task, and the
parent task of the sub-task is of a synchronized type, it link it to it.
If the task has an epic, and the epics are synchronized, it link it to
it. In last resort, the worklog will be linked with the project but with
no task.