Read access tokens with sudo
A normal user must be able to use the jira rest client under the hood.

Update documentation

Store original jira issues on analytic lines

Instead of the issue they are linked to in Odoo.
With the mechanism in place, if we don't import the task or bug issue
types and we synchronize the epics, the worklogs will be attached to the
epics of their tasks (of subtasks to tasks). The fields were showing the
values of the Epic (or task for subtask), though it makes much more
sense to keep the keys and URLs of the original issue and epic on the
analytic lines. We still have the link to the task if we want to get the
URL for the task they are currently linked to.

Fix project binding unique constraint

We should be able to have 2 bindings without external_id (not yet
exported)

Add wizard to link a task to JIRA

Refactorize multistep wizards with a mixin

Handle jira bindings in tasks merge wizard

Import name of Epics instead of summary

The Epic issues have a special (custom) field for their name.
In fact they use both, the custom field AND the summary field.
But I guess they are better identified by their epic name than
their summary.

Extract multi_step_wizard as an addon

Set a timeout on requests to JIRA

Fix typo in readme

Fix backend selection on wizard to link a task

The computed field was not called, due to the model being a
TransientModel, replace it by default values.

Add is_master on jira project bindings

Allow to have one project binding per project type

The unicity constraint (backend_id, odoo_id) on jira.backend.backend is
relaxed: it now allows one binding of each type.
The reason for this is:

* supporting several projects of different types is a requirements (eg.
  1 service desk and 1 software)
* but if we implement new features like "if I create a task it is
  pushed to Jira", with different projects we would not know where to
  push them

Using this constraint, we'll be able to focus new export features by
project type.

Move jira_key from project to jira project binding

As we can have more than one project binding, we cannot store a single
jira key. In case we have more than one jira key for a project, we
compute a compound key joined by commas.

Add method to handle jira API / connection errors

Add migration for 11.0.1.1.0

Fix error when several odoo users match

When we are importing a worklog for a user not yet linked and we found
several candidate users in odoo, they will both be linked with the same
jira user and make the unicity constraint fail.

Properly raise an error in this case.

Fix project_type as string

Add fallback project for worklogs

A new optional field on the backend allows to choose a fallback project
for the worklogs. When a worklog doesn't match any project linked with
Jira, they will be created there, allowing to find the misconfigurations
and fix them.

Add action to reimport worklogs from Jira

This is meant mainly to be used when a worklog has been imported in the
fallback project and we need to re-affect them to the correct project
after we linked it.

Add related action to open jira with import issues

The 'related' button on jobs which import issues or worklog will now
open jira directly on the issue.

Skip import of tasks/worklogs before batch dates

Jira may send webhooks for old records, for instance worklogs from 6
month ago because their task has been changed. This change ensures
that we never import any record which has a last update date before the
last date of batch import.

The job methods now return the result of Importer.run() so the result is
shown on the jobs in the UI.

Extract import of dependencies

Allowing to override partially the dependencies to import

Fix import of subtasks

When importing the parent of a subtask, the record in the 'parent' field
(the task data) is incomplete, it contains only a few fields. Providing
only the id to the dependency Importer will force it to read the whole
record from Jira.

Prevent duplicates with inactive projects

The constraint did not look for inactive projects, which
makes the import fail later because it finds several projects
for the same task.

Handle 404 errors when importing records

When a record does not exist on Jira:

* the job is done instead of failed
* a result on the job tells about the missing record
* the binding is deleted on Odoo
* for worklogs, the analytic line is deleted as well

Change xmlid for model_account_analytic_line

account is not necessarily installed, while analytic is installed by the
project module in any case

Partial revert of 2fa7d39

Commit title was: Skip import of tasks/worklogs before batch dates

The process is the following:
* Read T from backend, T is the is the last time we ran the import
* Create a batch job to import tasks/worklogs from T to now() and update
the backend with now()

The batch job is run asynchronously, it generates one job per task or
worklog to import between T and now(). When the jobs are executed, the
last batch date has already been updated to now(), so the jobs would
never import any record.

Fix error when force_reimport called on several records

Prevent changing task's project on binding

When the task Mapper binds a task to a task already existing in Odoo,
we must not change the project_id of the task. This is not possible as
long as we have invoiced timesheet lines, even if the id is the same
than the current one.

Add base for tests and a test for oauth

Add test for check of connection

With a fix in the method: getting the client only wasn't triggering any
error, calling 'myself' raises a 401 when we are not authenticated.

Add documentation for tests

Add travis configuration

Add tests on issue types and tasks

Add first test for analytic account lines

Make pylint-odoo happy

Add a script to initiate the Oauth dance