Fix tempo api call with actual username instead of key
The Jira API handles 2 ways to work with users: using the username or
the key field (but planning to migrate to anonymous accountIds).
Most of the time, the key is equal to the login, however if at some
point a login is changed, the key will always stay the same as the
original one. The Jira API handles both username and key field, but
Tempo accepts only the username.