Prevent error when linking users
When another user is already linked with the same jira user and matches.
Without this check, the transaction fails with a unique constraint
error.

Do not change project_id on tasks updates

When a task has invoiced timesheet lines, writing on 'project_id', even
if the id is the same, would fail because of a constraint preventing to
write project_id on timesheet lines.

Refactorize aal batch importer with 'updating since' jira method

This Jira method returns the list of ids updated since a timestamp, and
the youngest timestamp, which means we can store this timestamp and
reuse it the next time we ask the modified worklogs: this way we should
never miss any update.

To support this, a type of field which support milliseconds in datetime
is required, otherwise lose this precision that Jira gives us.

As of today, the "updated since" method exists on for worklogs on the
Jira webservice, so the rest of the batch import still have to use the
JQL filters (with minute-precision only).

Use Datetime as main format for MilliDatetime field

Store jira's 'updated_at' field to compare with binding

We can compare the millisecond dates on both sides to see if any changes
has been done on Jira and avoid useless writes.  Previously, we were
using the 'sync_date' field, which works too but will make more sync
than necessary (because the sync happens after the last jira's
updated_at value).

Fortuitously, the updated_at is also very useful for debugging purposes.

Add batch deleter for worklogs

Force reimport of worklogs as admin

Add tests for jira.backend.timestamp

Add jira.backend tests

Fix JiraTransactionCase → JiraSavepointCase 👀

Add tests for batch importer of tasks and worklogs

Add tests for worklogs deletes

Fix lint