Make migration more robust
It seems that when updating other addons, even if we don't update
connector_jira, it may already create the columns, then when we update
connector_jira, it tries to create the columns when they already exist.
Check before.

Use the rendered task description as HTML

The Jira REST API returns a rendered version of the description instead
of the markup when we pass the 'expand=renderedFields' parameter.

Update vrcpy cassettes as we read expanded issues

With "expand=renderedFields"