Fix oldname field
Apply suggestions from @simahawk in code review

Add scheduled action for importing deleted worklogs

Fix fetch pagination of organizations

The pagination specs of the servicedesk API are not the same than the
Jira API (respectively isLastPage vs lastPage). So the lib's
_fetch_pages does not fetch all the pages.

Optimize the requests: we get all the data at the initial call, so
the records are passed directly to "import_record" which will not do
another request.