Don't use the webhook payload for worklog
Partial revert of 13ba6ec

When a user records a tempo worklog, a Tempo Worklog is generated,
which also generates a Jira Worklog. This Jira Worklog is the one we
receive through the webhook. It is linked to the Tempo App user and not
to the Jira user who recorded the worklog, and the description is a
standard description.

-> if we want things to work when Tempo is used, we need to only use the
webhook as a notification, and not as a data source.